Revert to pre-Jack and Jill toolchain, add tests.
The only benefit to switching to Jack and Jill at this point in time
is using lambdas instead of anonymous classes. The downside is that
currently Mockito does not seem to be working with Jack. Thus, we will
revisit migration at a later time.

This commit adds some testing infrastructure (namely instrumented tests
using Espresso) that we can take advantage of going forward as we add
many more tests.