Add more tests and make "accounts" notion more robust.
We need to expand the current notion of an account to include who
provided the TOTP key (in our case that will be an exchange/wallet).
We will continue expanding upon this so that a user's Cryptodash
account can have any number of associated exchange/wallet accounts
linked with it. This logic will need to be added to the UI, which we
plan to do in as meaningful way as possible by using, for example,
the logos of each provider and adding data to the verify stage that
informs the user as to who they are linking their account with.